+++ game/ghdialogue/ghdview.py
@@ -27,8 +27,8 @@ class LancemateConvoItem(pbge.rpgmenu.MenuItem):
 class ConvoVisualizer(object):
     # The visualizer is a class used by the conversation when conversing.
     # It has a "text" property and "render", "get_menu" methods.
-    TEXT_AREA = pbge.frects.Frect(0,-125,350,150)
-    MENU_AREA = pbge.frects.Frect(0,50,350,204)
+    TEXT_AREA = pbge.frects.Frect(0,-125,350,180)
+    MENU_AREA = pbge.frects.Frect(0,80,350,174)
     NAME_AREA = pbge.frects.Frect(25,-185,300,35)
     REACT_AREA = pbge.frects.Frect(290,-185,35,35)
     PORTRAIT_AREA = pbge.frects.Frect(-370,-300,400,600)